refactor: smaller files and imrpoved structure
